+++ es6/modules/rawxml.js
@@ -1,6 +1,7 @@
 const traits = require("../traits");
 const { isContent } = require("../doc-utils");
 const { throwRawTagShouldBeOnlyTextInParagraph } = require("../errors");
+const { match, getValue } = require("../prefix-matcher");
 
 const moduleName = "rawxml";
 const wrapper = require("../module-wrapper");
@@ -52,27 +53,33 @@ function getInner({ part, left, right, postparsed, index }) {
 	return part;
 }
 
-const rawXmlModule = {
-	name: "RawXmlModule",
-	prefix: "@",
+class RawXmlModule {
+	constructor() {
+		this.name = "RawXmlModule";
+		this.prefix = "@";
+	}
 	optionsTransformer(options, docxtemplater) {
 		this.fileTypeConfig = docxtemplater.fileTypeConfig;
 		return options;
-	},
+	}
 	parse(placeHolderContent) {
 		const type = "placeholder";
-		if (placeHolderContent[0] !== this.prefix) {
-			return null;
+		if (match(this.prefix, placeHolderContent)) {
+			return {
+				type,
+				value: getValue(this.prefix, placeHolderContent),
+				module: moduleName,
+			};
 		}
-		return { type, value: placeHolderContent.substr(1), module: moduleName };
-	},
+		return null;
+	}
 	postparse(postparsed) {
 		return traits.expandToOne(postparsed, {
 			moduleName,
 			getInner,
 			expandTo: this.fileTypeConfig.tagRawXml,
 		});
-	},
+	}
 	render(part, options) {
 		if (part.module !== moduleName) {
 			return null;
@@ -85,7 +92,7 @@ const rawXmlModule = {
 			return { value: part.emptyValue || "" };
 		}
 		return { value };
-	},
+	}
 	resolve(part, options) {
 		if (!part.type === "placeholder" || part.module !== moduleName) {
 			return null;
@@ -98,7 +105,7 @@ const rawXmlModule = {
 				}
 				return value;
 			});
-	},
-};
+	}
+}
 
-module.exports = () => wrapper(rawXmlModule);
+module.exports = () => wrapper(new RawXmlModule());